Fix (new) potential dereference of NULL (top) pointer in writemsg(), reported by Deuce. If 'top' is passed in as NULL, assign a blank string constant to it since we dereference top a few places in this function and pass it to putmsg().

+++ sbbs3/msg_id.c
@@ -1,6 +1,6 @@
 /* Synchronet Message-ID generation routines */
 
-/* $Id: msg_id.c,v 1.14 2019/04/11 01:18:59 rswindell Exp $ */
+/* $Id: msg_id.c,v 1.13 2019/02/17 11:56:09 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -35,7 +35,7 @@
 
 #include "sbbs.h"
 
-static ulong msg_number(smbmsg_t* msg)
+static uint32_t msg_number(smbmsg_t* msg)
 {
 	if(msg->idx.number)
 		return(msg->idx.number);
@@ -56,7 +56,7 @@ uint32_t get_new_msg_number(smb_t* smb)
 	return smb->status.last_msg + 1;
 }
 
-static ulong msg_time(smbmsg_t* msg)
+static uint32_t msg_time(smbmsg_t* msg)
 {
 	if(msg->idx.time)
 		return(msg->idx.time);
